Quick refactor to move the throttling further back in the stack into the fan out stage and try a different delivery model
Previously the throttling was sitting on the SSE stream, this basically reads slow client, but the upstream fan out was as fast as upstream is, which in case of a client wanting to throttle meant we had slow consumers. Means more congestion, and higher overall latency as all our channel buffers slowly fill up.

Another issue was the delivery model, under load I would drop the latest events due to a filled buffer, now I'm trying to drain when I block and attempt an "enqueue latest". This is a knob for now, actually both are knobs just to test things out.